Fix some minor stuff
