+++ index.js
@@ -1,37 +1,29 @@
-'use strict';
+let express = require('express');
+let app = express();
+const port = process.env.PORT || 6000;
+const models = require('./models');
 
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
-const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/config/config.json')[env];
-const db = {};
+app.use(express.static(__dirname + '/public'));
 
-let sequelize;
-if (config.use_env_variable) {
-  sequelize = new Sequelize(process.env[config.use_env_variable], config);
-} else {
-  sequelize = new Sequelize(config.database, config.username, config.password, config);
-}
 
-fs
-  .readdirSync(__dirname)
-  .filter(file => {
-    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
-  })
-  .forEach(file => {
-    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
-    db[model.name] = model;
-  });
 
-Object.keys(db).forEach(modelName => {
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
+
+
+
+
+// Routing here
+app.use('/', require('./routes/indexRouter'));
+
+
+// listen log
+app.listen(port, () => {
+    console.log(`Server running at https://localhost:${port}`);
 });
 
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
-
-module.exports = db;
+// sync database
+app.get('/sync', (req, res) => {
+    models.sequelize.sync()
+    .then(() => {
+        res.send('Database synced completed');
+    });
+});
