+++ routes/indexRouter.js
@@ -1,0 +1,9 @@
+var express = require('express');
+var router = express.Router();
+var models = require('../models')
+
+router.get('/', (req, res) => {
+   res.render('index');
+})
+
+module.exports = router;
